05_multi_query_rag/requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.7
qdrant-client>=1.8.0
openai>=1.30.0
python-dotenv
